Remove percent encoding
Checked with the Android team. Either there is a real need (and an example to add to the unit tests) and we need a proper URL builder to avoid scattering encoding logic throughout the implementation, or there is none and we can simply forget about such issues. According to the Android team, no change was made to uids recently.